Revert "Revert "Revert "Contructor propery promotion, improved config and remove useless comments"""
