+++ app/Discord/Core/Builders/EmbedFactory.php
@@ -6,6 +6,9 @@ use App\Discord\Core\SlashCommand;
 use Discord\Builders\MessageBuilder;
 use Exception;
 
+/**
+ * Includes defined presets for easy usage throughout the code.
+ */
 class EmbedFactory
 {
     /**
