+++ app/Domain/Discord/Guild.php
@@ -12,13 +12,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
-use Illuminate\Support\Collection;
 
-/**
- * @property Collection $settings
- * @property Collection $channels
- * @property Collection $logSettings
- */
 class Guild extends Model
 {
     use HasFactory;
