+++ lang/en/bot.php
@@ -150,8 +150,6 @@ return [
         'add-level-msg' => 'Add custom level message',
         'add-welcome-msg' => 'Add custom welcome message',
         'guilds' => 'Overview of all guilds',
-        'vote' => 'Vote for Berend Botje on top.gg',
-        'support' => 'Support Server for Berend Botje',
     ],
 
     'msg' => [
